feat: add devcontainer
This is the start of a fully contained development environment.

The goal is to build and push a Docker image that contains a
complete development environment. The only dependency would be
Docker. Then with a command like...

```
$ docker run url-to-image
...
...
...
Open a browser to http://localhost:35000/
```

Then the user does all their work inside that browser.

Co-authored-by: Grant Braught <braught@dickinson.edu>